export_pdf/requirements.txt
psycopg2-binary
python-dotenv
matplotlib
reportlab
pandas
click